Converting test suite to use JUnit 5
Also uses ml-gradle to deploy a test application instead of depending on Documents and admin/admin.

Note that no application code is modified by this - the changes are all under src/test plus the addition of ml-gradle to deploy a test application along with some "test" scope additions to the pom.xml file. 

Also, I'm still getting intermittent timeout failures on ApplyTransformMarkLogicIT and QueryMarkLogicIT::testStructuredXMLQuery. That is the case on the ml-develop branch as well though.